Phase 6.0: pin ruff and fix lint drift, refresh CLAUDE.md
Pins ruff to an exact version and adds an explicit rule selection in
pyproject.toml so CI stops silently drifting as ruff's own default rule
set changes across releases (main was failing `ruff check .` with 32
findings even though no source behavior had changed).

Fixes the resulting findings without changing behavior: safe import/
pyupgrade autofixes, redundant int(round(...)) casts, an implicit
string concatenation, a B023 loop-variable closure, a naive strptime()
replaced with direct field parsing (decoder clock strings are elapsed
time, not calendar time), and documented noqa suppressions where the
flagged pattern is intentional (blind except for per-item batch
failure isolation, ValueError over TypeError to match the CLI's
existing exception contract).

Rewrites CLAUDE.md, which claimed the repository only implemented
Milestone 1 despite Phases 1-5.2 being complete, with the actual
current state and the architecture/scope principles the Phase 6 RF
survey work must follow.

All 50 existing tests pass unmodified; `ruff check .` is clean.

+++ src/dmr_iq_surveyor/spectrum/runner.py
@@ -196,7 +196,7 @@ def _analyze(
     waterfall_frequency_hz: np.ndarray | None = None
 
     bins_per_noise_window = max(
-        1, int(round(settings.local_noise_window_hz / resolution_hz))
+        1, round(settings.local_noise_window_hz / resolution_hz)
     )
     starts = iter_fft_starts(info.frame_count, settings.fft_size, settings.overlap_ratio)
 
@@ -524,7 +524,7 @@ def run_spectrum_batch(config_path: str | Path) -> dict[str, Any]:
                 settings=settings,
                 assumed_iq_order=iq_order,
             )
-        except Exception as exc:
+        except Exception as exc:  # noqa: BLE001 (one recording's failure must not abort the batch)
             rows.append(
                 {
                     "recording_id": recording.recording_id,
